data/code/requirements.txt
pre-commit
yapf == 0.26.0
flake8
pyyaml >= 5.1
visualdl >= 2.0.0

albumentations==0.5.1
numpy==1.18.1
opencv-python==4.2.0.32
p-tqdm==1.3.3
Pillow==7.0.0
scikit-image==0.17.2
six==1.14.0
tqdm==4.43.0